+++ assets_cleaner.yaml
@@ -9,5 +9,6 @@ config:
   # If you want to exclude specific
   # file from assets
   exclude-file:
-  #  - /assets/images/banner_upgrade.png # Uncomment this if you want to use
-  #  - /assets/images/logo.png # Uncomment this if you want to use
+    #  - /assets/images/banner_upgrade.png # Uncomment this if you want to use
+    #  - /assets/images/logo.png # Uncomment this if you want to use
+    - /assets/fonts/times-new-roman.ttf
